requirements.txt
appdirs==1.4.3
astroid==1.5.3
autopep8==1.3.4
beautifulsoup4==4.5.1
coverage==4.3.4
defusedxml==0.5.0
Django==1.10.3
django-allauth==0.30.0
django-common-helpers==0.9.0
django-cors-headers==2.0.2
django-cron==0.5.0
django-filter==1.0.1
django-nose==1.4.4
django-rest-auth==0.9.0
djangorestframework==3.5.3
django-rest-swagger==2.1.2
future==0.16.0
isort==4.2.5
lazy-object-proxy==1.2.2
lxml==3.7.2
Markdown==2.6.7
mccabe==0.6.1
nose==1.3.7
oauthlib==2.0.1
packaging==16.8
psycopg2==2.6.2
py==1.4.32
pycodestyle==2.3.1
pylint==1.7.2
pylint-django==0.7.2
pylint-plugin-utils==0.2.4
pyparsing==2.1.10
pytest==3.0.6
python-coveralls==2.9.1
python3-openid==3.1.0
PyYAML==3.12
requests==2.12.1
requests-oauthlib==0.8.0
six==1.10.0
wrapt==1.10.8
gunicorn==19.7.0
raven==6.1.0
